# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express API with Swagger",
    "version": "1.0.0"
  },
  "paths": {
...
    "/duty/publish": {
      "get": {
        "summary": "Display the duty publishing form with specialties",
        "operationId": "getDutyPublishForm",
        "security": [
          {
            "sessionAuth": []
          },
          {
            "hospitalRole": []
          }
        ],
        "responses": {
          "200": {
            "description": "Duty publishing form page rendered successfully",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error occurred while fetching specialties",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "securitySchemes": {
        "sessionAuth": {
          "type": "apiKey",
          "in": "cookie",
          "name": "sessionToken",
          "description": "Session token used for authenticating API requests"
        },
        "hospitalRole": {
          "type": "apiKey",
          "in": "header",
          "name": "Role",
          "description": "Ensure the user role is 'hospital'"
        },
        "doctorRole": {
          "type": "apiKey",
          "in": "header",
          "name": "Role",
          "description": "Ensure the user role is 'doctor'"
        }
      },
      "post": {
        "summary": "Publish a new duty slot",
        "operationId": "publishDutySlot",
        "security": [
          {
            "sessionAuth": []
          },
          {
            "hospitalRole": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishDutySlotRequest"
              }
            }
          }
        },
        "responses": {
          "303": {
            "description": "Redirects after publishing duty slot",
            "headers": {
              "Location": {
                "description": "URL to redirect to after successful publishing",
                "schema": {
                  "type": "string",
                  "example": "/"
                }
              }
            }
          },
          "500": {
            "description": "Error occurred while publishing duty slot",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/duty/find_by_specialty": {
      "get": {
        "summary": "Find duty slots by specialty",
        "operationId": "findDutyBySpecialty",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "specialty",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Specialty to filter duty slots"
          }
        ],
        "responses": {
          "200": {
            "description": "List of duty slots filtered by specialty",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "dutySlots": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DutySlot"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid specialty provided",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error fetching duty slots",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/duty/remove": {
      "post": {
        "summary": "Remove a duty slot by ID",
        "operationId": "removeDutySlot",
        "security": [
          {
            "sessionAuth": []
          },
          {
            "hospitalRole": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "description": "ID of the duty slot to remove"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "303": {
            "description": "Redirects after removing duty slot",
            "headers": {
              "Location": {
                "description": "URL to redirect to after successful removal",
                "schema": {
                  "type": "string",
                  "example": "/"
                }
              }
            }
          },
          "404": {
            "description": "Duty slot not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error occurred while removing duty slot",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/duty/slots/json": {
      "get": {
        "summary": "Retrieve all duty slots",
        "operationId": "getDutySlots",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of duty slots",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DutySlot"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/assign-duty-slot": {
      "post": {
        "summary": "Assign a duty slot to a doctor",
        "operationId": "assignDutySlot",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDutySlotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Duty slot updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DutySlotResponse"
                }
              }
            }
          },
          "404": {
            "description": "Duty slot not found or already assigned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/give-consent": {
      "post": {
        "summary": "Give consent for a duty slot",
        "operationId": "giveConsent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "description": "ID of the duty slot"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consent given successfully, duty slot updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DutySlotResponse"
                }
              }
            }
          },
          "404": {
            "description": "Duty slot not found or not in pending status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/revoke-assignment": {
      "post": {
        "summary": "Revoke an assigned duty slot",
        "operationId": "revokeAssignment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "description": "ID of the duty slot"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assignment revoked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DutySlotResponse"
                }
              }
            }
          },
          "404": {
            "description": "Duty slot not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Specialty": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "PublishDutySlotRequest": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "dutyHours": {
            "type": "string"
          },
          "requiredSpecialty": {
            "type": "string",
            "description": "ID of the required specialty"
          }
        }
      },
      "DutySlot": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "hospitalId": {
            "type": "string",
            "description": "ID of the hospital that published the duty slot",
            "$ref": "#/components/schemas/User"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "dutyHours": {
            "type": "string"
          },
          "requiredSpecialty": {
            "type": "string",
            "description": "ID of the required specialty",
            "$ref": "#/components/schemas/Specialty"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "pending",
              "filled"
            ]
          },
          "assignedDoctorId": {
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ],
            "nullable": true
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "doctor",
              "hospital"
            ]
          },
          "specialty": {
            "type": "string",
            "nullable": true
          },
          "localization": {
            "type": "string",
            "nullable": true
          },
          "profileVisible": {
            "type": "boolean"
          }
        }
      },
      "AssignDutySlotRequest": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "sendingDoctorId": {
            "type": "string"
          }
        }
      },
      "DutySlotResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "slot": {
            "$ref": "#/components/schemas/DutySlot"
          }
        }
      },
...
    },
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionToken",
        "description": "Session token used for authenticating API requests"
      },
      "doctorRole": {
        "type": "apiKey",
        "in": "header",
        "name": "Role",
        "description": "Ensure the user role is 'doctor'"
      }
    }
  },
...
}